% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondirtree.R
\name{dir2json}
\alias{dir2json}
\title{JSON representation of a directory}
\usage{
dir2json(dir, depth = NULL)
}
\arguments{
\item{dir}{the path of the directory}

\item{depth}{a nonnegative integer, the desired depth, or \code{NULL} (default) to search until the end}
}
\value{
A string.
}
\description{
Returns the hierarchical structure of a folder as a JSON string.
}
\examples{
dir2json(".", 0)
dir2json(".", 1)
dir2json(Sys.getenv("R_HOME"), 1)
}
