% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{json_decode_dir}
\alias{json_decode_dir}
\title{Decode JSON to a Directory}
\usage{
json_decode_dir(json_data, dir)
}
\arguments{
\item{json_data}{A JSON string representing the directory's contents.}

\item{dir}{A character string specifying the target directory.}
}
\value{
None. Creates files in the specified directory.
}
\description{
This function decodes a JSON string into a directory structure.
}
