% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondirtree.R
\name{shinyDataTree}
\alias{shinyDataTree}
\title{Shiny app: tree representation of a dataframe}
\usage{
shinyDataTree(dat, root = TRUE)
}
\arguments{
\item{dat}{the data to be represented}

\item{root}{Boolean, whether to include a root node}
}
\value{
No value, just runs a Shiny app.
}
\description{
Shiny app to play with a Reingold-Tilford tree network diagram
  representing a hierarchical data structure.
}
\examples{
library(dir2json)
dat <- tibble::tribble(
  ~level1, ~level2, ~level3, ~level4, 
  "Beverages", "Water", "", "", 
  "Beverages", "Coffee", "", "", 
  "Beverages", "Tea", "Black tea", "", 
  "Beverages", "Tea", "White tea", "", 
  "Beverages", "Tea", "Green tea", "Sencha", 
  "Beverages", "Tea", "Green tea", "Gyokuro", 
  "Beverages", "Tea", "Green tea", "Matcha", 
  "Beverages", "Tea", "Green tea", "Pi Lo Chun"
)
shinyDataTree(dat)
}
