% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondirtree.R
\name{shinyDirTree}
\alias{shinyDirTree}
\title{Shiny app: tree representation of a folder}
\usage{
shinyDirTree(dir = NULL, depth = NULL)
}
\arguments{
\item{dir}{path to a directory; if \code{NULL}, it is selected in the app}

\item{depth}{the desired depth, ignored if \code{dir=NULL}; see \code{\link{dir2json}}}
}
\description{
Shiny app to play with a Reingold-Tilford tree network diagram
representing a folder structure
}
