% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondirtree.R
\name{shinySheetTree}
\alias{shinySheetTree}
\title{Shiny app: tree representation of an Excel table}
\usage{
shinySheetTree(xlsx, sheet = 1, root = TRUE)
}
\arguments{
\item{xlsx}{path to a xlsx file}

\item{sheet}{sheet to read; either a string (the name of a sheet), or an 
integer (the position of the sheet)}

\item{root}{Boolean, whether to include a root node}
}
\value{
No value, just runs a Shiny app.
}
\description{
Shiny app to play with a Reingold-Tilford tree network diagram
  representing a hierarchical data structure.
}
\note{
This function runs \code{\link{shinyDataTree}} with the dataframe 
  imported from the xlsx file.
}
