% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{text_file_extensions}
\alias{text_file_extensions}
\title{Get Text File Extensions}
\usage{
text_file_extensions()
}
\value{
A character vector of text file extensions.
}
\description{
This function returns a vector of common text file extensions.
}
