% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{json_encode_dir}
\alias{json_encode_dir}
\title{Encode a Directory to JSON}
\usage{
json_encode_dir(dir)
}
\arguments{
\item{dir}{A character string specifying the directory to encode.}
}
\value{
A JSON string representing the directory's contents.
}
\description{
This function encodes all files in a directory into a JSON format.
}
